+++ test/couchdb_test.dart
@@ -1,7 +1,7 @@
 import 'package:couchdb/couchdb.dart';
 
 Future<void> main() async {
-  final c = CouchDbClient(username: 'name', password: 'password');
+  final c = CouchDbClient.fromString('http://name:password@localhost:5984');
   final da = DatabaseModel(c);
   final ddm = DesignDocumentModel(c);
   final dm = DocumentModel(c);
@@ -11,9 +11,9 @@ Future<void> main() async {
 
   try {
     final headers = <String, String>{'Accept': 'text/plain'};
-    
+
     //final r = await c.authenticate();
-    
+
     final o = await da.shards('denta');
     print(o.databaseModelResponse().shards);
   } on CouchDbException catch (e) {